requirements.txt
streamlit
bcrypt
requests
python-dotenv
